+++ comics/comics/eatthattoast.py
@@ -9,19 +9,11 @@ class ComicData(ComicDataBase):
     language = 'en'
     url = 'http://eatthattoast.com/'
     start_date = '2010-06-14'
+    end_date = '2013-12-02'
+    active = False
     rights = 'Matt Czapiewski'
 
 
 class Crawler(CrawlerBase):
-    history_capable_days = 90
-    time_zone = 'US/Eastern'
-
     def crawl(self, pub_date):
-        page = self.parse_page('http://eatthattoast.com/')
-        url = page.src('#comic img')
-        if url:
-            title = page.text('.comicpress_comic_title_widget a')
-            text = page.alt('#comic img')
-            matches = re.match(r'.*(\d{4}-\d{2}-\d{2}).*', url)
-            if matches and matches.groups()[0] == pub_date.isoformat():
-                return CrawlerImage(url, title, text)
+        pass  # Comic no longer published